changed search courses endpoint
