+++ backend/app/main.py
@@ -1,6 +1,6 @@
 from fastapi import Depends, FastAPI
 
-from .routers import users
+from .routers import *
 
 app = FastAPI(
     docs_url="/swagger",      # вместо /docs
@@ -9,6 +9,7 @@ app = FastAPI(
 )
 
 app.include_router(users.router)
+app.include_router(courses.router)
 
 @app.get("/")
 async def root():
