+++ backend/app/routers/__init__.py
@@ -1,0 +1,2 @@
+from .users import router
+from .courses import router
